moving Auth Providers components to store-ui
