+++ packages/gatsby-theme-store/src/components/Auth/Providers/EmailVerification/index.tsx
@@ -1,14 +1,16 @@
 import React, { useReducer } from 'react'
 import type { FC } from 'react'
+import {
+  EmailVerificationAccessCodeForm as AccessCodeForm,
+  EmailVerificationEmailForm as EmailForm,
+  EmailVerificationReducer as reducer,
+} from '@vtex/store-ui'
+import type { AuthProviderComponentProps } from '@vtex/store-ui'
 
 import { sendAccessKey, validateAccessKey } from '../../../../sdk/auth/Service'
 import { useOnLoginSuccessful } from '../../../../sdk/auth/useOnLoginSuccessful'
 import { useStartLogin } from '../../../../sdk/auth/useStartLogin'
 import { isValidAccessCode, isValidEmail } from '../../../../sdk/auth/validate'
-import AccessCodeForm from './AccessCodeForm'
-import EmailForm from './EmailForm'
-import { reducer } from './state'
-import type { AuthProviderComponentProps } from '../types'
 
 const EmailVerification: FC<AuthProviderComponentProps> = ({
   variant: v,
