+++ packages/store-ui/src/Auth/Providers/EmailVerification/Button.tsx
@@ -1,5 +1,5 @@
 import { useIntl } from '@vtex/gatsby-plugin-i18n'
-import { Button } from '@vtex/store-ui'
+import { Button } from 'theme-ui'
 import React from 'react'
 import type { FC } from 'react'
