+++ packages/store-ui/src/Auth/Providers/index.tsx
@@ -1,0 +1,27 @@
+// EmailAndPassword
+export { default as EmailAndPasswordButton } from './EmailAndPassword/Button'
+export { default as EmailAndPasswordEmailForm } from './EmailAndPassword/EmailForm'
+export { default as EmailAndPasswordSignInForm } from './EmailAndPassword/SignInForm'
+export { default as EmailAndPasswordSignUpForm } from './EmailAndPassword/SignUpForm'
+export {
+  State as EmailAndPasswordState,
+  Action as EmailAndPasswordAction,
+  reducer as EmailAndPasswordReducer,
+} from './EmailAndPassword/state'
+// EmailVerification
+export { default as EmailVerificationAccessCodeForm } from './EmailVerification/AccessCodeForm'
+export { default as EmailVerificationButton } from './EmailVerification/Button'
+export { default as EmailVerificationEmailForm } from './EmailVerification/EmailForm'
+export {
+  State as EmailVerificationState,
+  Action as EmailVerificationAction,
+  reducer as EmailVerificationReducer,
+} from './EmailVerification/state'
+// Facebook
+export { default as FacebookOAuthButton } from './Facebook/Button'
+// Google
+export { default as GoogleOAuthButton } from './Google/Button'
+// Theme
+export { default as AuthProvidersTheme } from './theme'
+// Types
+export * from './types'
