+++ packages/store-ui/src/index.tsx
@@ -113,3 +113,5 @@ export { default as CarouselPage } from './Carousel/Page'
 export { default as CarouselArrowLeft } from './Carousel/ArrowLeft'
 export { default as CarouselArrowRight } from './Carousel/ArrowRight'
 export { default as CarouselPaginationDots } from './Carousel/PaginationDots'
+// Auth Providers
+export * from './Auth/Providers'
